refactor(voip): extract inbound invite router
